Moves HashCoordinate trait to geometry traits module
Relocates the `HashCoordinate` trait from the point module to the geometry traits module.

This change improves code organization and promotes better separation of concerns by grouping traits together.

This also includes adding tests for trait implementations.
